Feature/final fix of bugs

+++ packages/App/src/balance/wallet/Wallet.tsx
@@ -17,16 +17,14 @@ import {
   addNewCardError,
   filteredIncome,
   filteredExp,
+  getError,
 } from '../../store/selectors/walletItems';
 import {useDispatch, useSelector} from 'react-redux';
 import {
-  filterInComeRequest,
-  filterExpensesRequest,
   deleteWalletCardRequest,
   cardMonetaryMove,
   deleteTransactionRequest,
   correctTransactionInfo,
-  filterAllItemsRequest,
 } from '../../store/actions/walletActions';
 import Loading from '../../components/Loading';
 import {
@@ -34,11 +32,13 @@ import {
   WalletInfo,
   ITransactions,
 } from '../../types/types';
-import CardModal from '../../components/CardModal';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
+import {useDeviceOrientation} from '@react-native-community/hooks/lib/useDeviceOrientation';
+import NetInfo from '@react-native-community/netinfo';
+import dayjs from 'dayjs';
 
 const viewability: ViewabilityConfig = {
-  viewAreaCoveragePercentThreshold: 50,
+  viewAreaCoveragePercentThreshold: 10,
 };
 
 const keyExtractorForCards = (item: WalletInfo) => String(item.id);
@@ -51,7 +51,7 @@ interface Props {
 
 const Wallet: FC<Props> = ({navigation}) => {
   const dispatch = useDispatch();
-
+  const orientation = useDeviceOrientation();
   const receivedWalletItems = useSelector(walletItems);
   const filterIncome = useSelector(filteredIncome);
   const filteredExpenses = useSelector(filteredExp);
@@ -60,44 +60,50 @@ const Wallet: FC<Props> = ({navigation}) => {
   const errorFilters = useSelector(filtersError);
   const errorDeleteCard = useSelector(deleteCardError);
   const errorAddCard = useSelector(addNewCardError);
+  const getErrorInfo = useSelector(getError);
 
   const [itemVisible, setItemVisible] = useState<number>(0);
-  const [isModalCardVisible, setIsModalCardVisible] = useState<boolean>(false);
-  const [isModalTransactionVisible, setIsModalTransactionVisible] =
-    useState<boolean>(false);
-  const [cardId, setCardId] = useState<number>(0);
-  const [transactionKey, setTransactionKey] = useState<{
-    keyTransaction: number;
-    amount: number;
-    type: string;
-  }>({keyTransaction: 0, amount: 0, type: ''});
   const [refItem, setRefItem] = useState<FlatList<WalletInfo> | null>();
   const [chosenBtn, setChosenBtn] = useState<string>('All actions');
-  const [categoryShows, setCategoryShows] =
-    useState<Array<WalletInfo>>(receivedWalletItems);
+  const [isConnection, setIsConnection] = useState<boolean>(true);
+  const [lastTime, setLastTime] = useState<Date>();
+
+  useEffect(() => {
+    setIsConnection(true);
+    NetInfo.addEventListener(state => {
+      console.log(state.isConnected);
+      if (state.isConnected === false) {
+        setIsConnection(false);
+      }
+      if (state.isConnected === true) {
+        setIsConnection(true);
+        setLastTime(new Date());
+      }
+    });
+  }, []);
+
+  if (isConnection === false) {
+    if (lastTime) {
+      Alert.alert(
+        'No internet',
+        `Last time connection was ${dayjs(lastTime).format('DD MMM, hh:mm')}`,
+      );
+    }
+    if (!lastTime) {
+      Alert.alert('No internet connection');
+    }
+  }
 
   const filterInCome = (title: string) => {
     setChosenBtn(title);
-    dispatch(filterInComeRequest());
-    if (filterIncome[itemVisible].transactions) {
-      setCategoryShows(filterIncome);
-    }
   };
 
   const filterExpenses = (title: string) => {
     setChosenBtn(title);
-    dispatch(filterExpensesRequest());
-    if (filteredExpenses[itemVisible].transactions) {
-      setCategoryShows(filteredExpenses);
-    }
   };
 
   const allCategories = (title: string) => {
     setChosenBtn(title);
-    dispatch(filterAllItemsRequest());
-    if (receivedWalletItems[itemVisible].transactions) {
-      setCategoryShows(receivedWalletItems);
-    }
   };
 
   const newCard = () => {
@@ -114,15 +120,21 @@ const Wallet: FC<Props> = ({navigation}) => {
   };
 
   const showModal = (id: number) => {
-    setCardId(id);
-    setIsModalCardVisible(true);
-  };
-
-  const deleteCard = () => {
-    dispatch(deleteWalletCardRequest(cardId));
-    setItemVisible(0);
-    refItem?.scrollToIndex({animated: true, index: 0});
-    setIsModalCardVisible(false);
+    Alert.alert(' Would you like to delete card?', '', [
+      {
+        text: 'Cancel',
+        style: 'cancel',
+      },
+      {
+        text: 'Delete',
+        onPress: () => {
+          const indexCard = itemVisible - 1;
+          dispatch(deleteWalletCardRequest(id));
+          setItemVisible(0);
+          refItem?.scrollToIndex({animated: true, index: indexCard});
+        },
+      },
+    ]);
   };
 
   const showModalTransaction = (
@@ -130,14 +142,20 @@ const Wallet: FC<Props> = ({navigation}) => {
     amount: number,
     type: string,
   ) => {
-    setTransactionKey({keyTransaction, amount, type});
-    setIsModalTransactionVisible(true);
-  };
-
-  const deleteTransaction = () => {
-    const item = receivedWalletItems[itemVisible];
-    dispatch(deleteTransactionRequest({item, transactionKey}));
-    setIsModalTransactionVisible(false);
+    const transactionKey = {keyTransaction, amount, type};
+    Alert.alert('Would you like to delete transaction?', '', [
+      {
+        text: 'Cancel',
+        style: 'cancel',
+      },
+      {
+        text: 'Delete',
+        onPress: () => {
+          const item = receivedWalletItems[itemVisible];
+          dispatch(deleteTransactionRequest({item, transactionKey}));
+        },
+      },
+    ]);
   };
 
   const correctTransaction = (
@@ -171,7 +189,10 @@ const Wallet: FC<Props> = ({navigation}) => {
     if (errorAddCard) {
       Alert.alert(errorAddCard);
     }
-  }, [errorDeleteCard, errorAddCard]);
+    if (getErrorInfo) {
+      Alert.alert(getErrorInfo);
+    }
+  }, [errorDeleteCard, errorAddCard, getErrorInfo]);
 
   if (isLoading) {
     <Loading />;
@@ -179,23 +200,10 @@ const Wallet: FC<Props> = ({navigation}) => {
 
   return (
     <View style={styles.container}>
-      <CardModal
-        title="Would you like to delete transaction?"
-        isVisible={isModalTransactionVisible}
-        onPressDelete={deleteTransaction}
-        onPressHide={setIsModalTransactionVisible}
-      />
-
-      <CardModal
-        title=" Would you like to delete card?"
-        isVisible={isModalCardVisible}
-        onPressDelete={deleteCard}
-        onPressHide={setIsModalCardVisible}
-      />
-
-      <View style={styles.wallet}>
-        <Text style={styles.title}>Total</Text>
-        <Text>{receivedSum}$</Text>
+      <View
+        style={orientation.landscape ? styles.walletLandscape : styles.wallet}>
+        <Text style={styles.title}>TOTAL BALANCE</Text>
+        <Text style={styles.titleAmount}>{receivedSum}$</Text>
       </View>
 
       {isLoading ? (
@@ -228,35 +236,34 @@ const Wallet: FC<Props> = ({navigation}) => {
                 />
               </View>
 
-              <View style={styles.buttonArea}>
-                <Button
-                  title="All actions"
-                  picture={allCategoriesIconSource}
-                  onPress={allCategories}
-                  chosen={chosenBtn}
-                />
-                <Button
-                  title="Incoming"
-                  picture={incomeIconSource}
-                  onPress={filterInCome}
-                  chosen={chosenBtn}
-                />
-                <Button
-                  title="Expenses"
-                  picture={expensesIconSource}
-                  onPress={filterExpenses}
-                  chosen={chosenBtn}
-                />
-
-                <Button
-                  title="New card"
-                  picture={addNewIconSource}
-                  onPress={newCard}
-                  chosen={chosenBtn}
-                />
-              </View>
-
               <View style={styles.categoriesListTitle}>
+                <View style={styles.buttonArea}>
+                  <Button
+                    title="All actions"
+                    picture={allCategoriesIconSource}
+                    onPress={allCategories}
+                    chosen={chosenBtn}
+                  />
+                  <Button
+                    title="Incoming"
+                    picture={incomeIconSource}
+                    onPress={filterInCome}
+                    chosen={chosenBtn}
+                  />
+                  <Button
+                    title="Expenses"
+                    picture={expensesIconSource}
+                    onPress={filterExpenses}
+                    chosen={chosenBtn}
+                  />
+
+                  <Button
+                    title="New card"
+                    picture={addNewIconSource}
+                    onPress={newCard}
+                    chosen={chosenBtn}
+                  />
+                </View>
                 <Text style={styles.categoriesListText}>All categories</Text>
               </View>
             </>
@@ -264,7 +271,9 @@ const Wallet: FC<Props> = ({navigation}) => {
           data={
             chosenBtn === 'All actions'
               ? receivedWalletItems[itemVisible].transactions
-              : categoryShows[itemVisible].transactions
+              : chosenBtn === 'Incoming'
+              ? filterIncome[itemVisible].transactions
+              : filteredExpenses[itemVisible].transactions
           }
           keyExtractor={keyExtractorForTransactions}
           renderItem={({item}) => (
@@ -289,36 +298,50 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     flexDirection: 'column',
-    backgroundColor: '#FFFFFF',
-    paddingBottom: 70,
+    backgroundColor: 'white',
   },
 
   list: {
-    height: 180,
+    height: 100,
   },
 
   wallet: {
+    alignItems: 'center',
+    width: '100%',
+    paddingVertical: 20,
+    backgroundColor: '#FFFFFF',
+  },
+
+  walletLandscape: {
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
     width: '100%',
-    height: 55,
-    paddingHorizontal: 20,
-    backgroundColor: '#D0EEFF',
+    paddingVertical: 10,
+    backgroundColor: '#FFFFFF',
   },
 
   title: {
-    fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: '700',
+    color: '#A8ADDD',
+    fontSize: 14,
+  },
+
+  titleAmount: {
+    fontWeight: '600',
+    fontSize: 32,
     color: 'black',
-    fontSize: 18,
   },
 
   buttonArea: {
+    backgroundColor: '#F6F6F6',
+    justifyContent: 'center',
+    alignItems: 'center',
     flexDirection: 'row',
-    margin: 20,
-    height: 70,
+    padding: 3,
+    borderRadius: 8,
+    marginTop: 20,
   },
   categoriesList: {
     flex: 1,
@@ -327,12 +350,19 @@ const styles = StyleSheet.create({
   },
 
   categoriesListTitle: {
-    justifyContent: 'center',
+    paddingTop: 10,
     alignItems: 'center',
+    backgroundColor: 'white',
+    borderTopLeftRadius: 10,
+    borderTopRightRadius: 10,
+    elevation: 10,
+  },
+
+  categoriesListTitleBack: {
+    backgroundColor: '#EDEFFE',
   },
 
   categoriesListText: {
-    fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: '800',
     fontSize: 18,
